CommentTest 보강/수정
